Перенёс используемые типы в services/types
