+++ src/components/card/card.tsx
@@ -3,8 +3,8 @@ import {useLocation, useNavigate} from "react-router-dom";
 import {useSelector} from "react-redux";
 import {useDrag} from "react-dnd";
 import {Counter, CurrencyIcon} from '@ya.praktikum/react-developer-burger-ui-components';
+import {IngredientType} from "../../services/types/data";
 import styles from './card.module.css';
-import {IngredientType} from "../../utils/types";
 
 interface ICardProps {
     item: IngredientType;
