+++ src/components/ingredient-details/ingredient-details.tsx
@@ -1,7 +1,7 @@
 import {useEffect, useState} from "react";
 import {useSelector} from "react-redux";
 import {useParams} from "react-router-dom";
-import {IngredientType} from "../../utils/types";
+import {IngredientType} from "../../services/types/data";
 import style from './ingredient-details.module.css';
 
 const IngredientDetails = () => {
