+++ src/services/types/data.ts
@@ -12,17 +12,3 @@ export type IngredientType = {
     image_large: string;
     __v: number;
 }
-
-export type SelectedIngredientType = IngredientType & {
-    id: string;
-}
-
-export type LoginType = {
-    email: string;
-    password: string;
-}
-
-export type UserType = {
-    name: string;
-} & LoginType;
-
